Update JSON schema
Change the formatting style with the 'pretty-format-json' pre-commit hook. Also apply changes that were flagged by the 'sourcemeta-jsonschema-lint' pre-commit hook. These include redundant schema definitions (e.g. setting 'type' alongside 'const' or 'enum') and correcting an inconsistency between the subschema definition for the 'Conjugate' property and the example data. Finally, add an '$id' property to the schema so that it can be correctly resolved and matched with test cases in the 'test/schema_tests/' test suite.